python_gen/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0,<1.0
pillow>=10.0.0,<11
pytest>=8.0.0,<9
pytest-asyncio>=0.24.0,<1
httpx>=0.27.0,<1
